+++ packages/expo-print/src/ExponentPrint.ts
@@ -1,2 +1,6 @@
-import { requireNativeModule } from 'expo-modules-core';
-export default requireNativeModule('ExpoPrint');
+import { requireNativeModule, NativeModulesProxy, Platform } from 'expo-modules-core';
+
+const printModule =
+  Platform.OS === 'ios' ? requireNativeModule('ExpoPrint') : NativeModulesProxy.ExponentPrint;
+
+export default printModule;